+++ src/main/webapp/WEB-INF/web.xml
@@ -40,7 +40,7 @@
 		
 	<servlet-mapping>
 		<servlet-name>appServlet</servlet-name>
-		<url-pattern>*.do</url-pattern>
+		<url-pattern>/</url-pattern>
 	</servlet-mapping>
 	<welcome-files>
 		<welcome-file>index.jsp</welcome-file>
